Remove companion from StreamFileProvider
